refactor: parser 폴더 정리
