+++ script/llm/parser/analyzer/variableAnalyzer.ts
