+++ script/llm/parser/type/interface.ts
